requirements.txt
mysql-connector-python~=8.0.28
requests~=2.26.0
pytest~=6.2.5
selenium~=4.1.0